Extract AixOperatingSystem base and shared perfstat process carrier
AIX was the only platform whose OperatingSystem JNA and FFM twins extended
AbstractOperatingSystem directly, with no common base, so they duplicated the
service/version/boot-time/installed-app logic and the perfstat-driven process
and thread enumeration.

Introduce a backend-neutral AixPerfstatProcess carrier (oshi-common) holding
the perfstat_process fields OSHI consumes; PerfstatProcessJNA maps its JNA
struct into it and PerfstatProcessFFM returns it in place of its private
ProcessInfo POJO. On top of that, extract an AixOperatingSystem base holding
all the shared logic plus the carrier-typed process/thread enumeration. The
twins keep only hooks: the perfstat/partition-config reads, the native
process/thread IDs, and the OSProcess/InternetProtocolStats/NetworkParams
factories.

AixOSProcess also unifies: its procCpu iteration (updateAttributes) and the
getrlimit open-file-limit skeleton move into the base behind an
AixPerfstatProcess[] supplier and a queryRlimitNofile hook, so the twins keep
only the native getrlimit, affinity, and address-space reads.

Behavior-preserving refactor; no CHANGELOG.
