+++ oshi-core-ffm/src/main/java/oshi/driver/unix/aix/perfstat/PerfstatProcessFFM.java
@@ -22,6 +22,7 @@ import org.slf4j.LoggerFactory;
 import org.slf4j.event.Level;
 
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.driver.common.unix.aix.AixPerfstatProcess;
 import oshi.ffm.ForeignFunctions;
 
 /**
@@ -35,17 +36,6 @@ public final class PerfstatProcessFFM {
     private PerfstatProcessFFM() {
     }
 
-    /** POJO mirror of the {@code perfstat_process_t} fields OSHI consumes. */
-    public static final class ProcessInfo {
-        public long pid;
-        public long num_threads;
-        public long proc_real_mem_data;
-        public long proc_real_mem_text;
-        public long real_inuse;
-        public double ucpu_time;
-        public double scpu_time;
-    }
-
     /** Slack added to the perfstat_process count to absorb new processes between count and fill calls. */
     private static final int PROC_COUNT_PAD = 10;
 
@@ -56,25 +46,25 @@ public final class PerfstatProcessFFM {
      * perfstat calls — mirrors {@code PerfstatProcessJNA} and the {@code MacOperatingSystemJNA.getThreadCount}
      * precedent (+10).
      *
-     * @return one {@link ProcessInfo} per process, or an empty array on error
+     * @return one {@link AixPerfstatProcess} per process, or an empty array on error
      */
-    public static ProcessInfo[] queryProcesses() {
+    public static AixPerfstatProcess[] queryProcesses() {
         return ForeignFunctions.callInArenaOrDefault(arena -> {
             int count = perfstat_process(MemorySegment.NULL, MemorySegment.NULL, PERFSTAT_PROCESS_T_SIZE, 0);
             if (count <= 0) {
-                return new ProcessInfo[0];
+                return new AixPerfstatProcess[0];
             }
             int padded = count + PROC_COUNT_PAD;
             MemorySegment buf = arena.allocate((long) PERFSTAT_PROCESS_T_SIZE * padded);
             MemorySegment firstName = arena.allocate(PERFSTAT_ID_T_SIZE);
             int ret = perfstat_process(firstName, buf, PERFSTAT_PROCESS_T_SIZE, padded);
             if (ret <= 0) {
-                return new ProcessInfo[0];
+                return new AixPerfstatProcess[0];
             }
-            ProcessInfo[] result = new ProcessInfo[ret];
+            AixPerfstatProcess[] result = new AixPerfstatProcess[ret];
             for (int i = 0; i < ret; i++) {
                 long off = (long) i * PERFSTAT_PROCESS_T_SIZE;
-                ProcessInfo p = new ProcessInfo();
+                AixPerfstatProcess p = new AixPerfstatProcess();
                 p.pid = procPid(buf, off);
                 p.num_threads = procNumThreads(buf, off);
                 p.proc_real_mem_data = procRealMemData(buf, off);
@@ -85,6 +75,6 @@ public final class PerfstatProcessFFM {
                 result[i] = p;
             }
             return result;
-        }, LOG, Level.TRACE, "Failed to query process statistics", new ProcessInfo[0]);
+        }, LOG, Level.TRACE, "Failed to query process statistics", new AixPerfstatProcess[0]);
     }
 }
